backend/requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
google-auth==2.28.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
oauthlib==2.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.1.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==2.3.7
